Refactor string formatting classes to improve structure and reuse
- Introduced NumberFormatterBase as a base class for numeric formatters, consolidating common formatting logic.
- Created StringFormatterBase for string formatters, providing basic width formatting functionality.
- Updated FloatFormatter, ExponentialFormatter, HexFloatFormatter, and others to extend from NumberFormatterBase.
- Refactored CharacterFormatter and PointerFormatter to extend from StringFormatterBase.
- Implemented DoubleUtils for handling IEEE 754 double-precision floating-point representation.
- Removed redundant code and improved clarity in formatting methods.
